requirements.txt
Flask
python-docx
docxtpl
docx2pdf
PyMuPDF
Pillow
gunicorn